3-rd commit - fixing statistic bugs and concurrent

+++ src/main/java/ru/cgm/testcgm/stat/QueriesStatistics.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class QueriesStatistics {
-    private int totalQueries;
-    private int averageInStrings;
-    private int averageMaxSequence;
+    private float totalQueries;
+    private float averageInStrings;
+    private float averageMaxSequence;
 }
